refactored spellSearchRouter to use typeorm to query the db
