+++ server/src/routes/spellSearch/index.ts
@@ -1,3 +1,4 @@
-export * from './processQuery';
-export * from './createSQLParameters';
-export * from './router';
+export * from './spellEntity';
+export * from './spellRouter';
+export * from './spellService';
+export * from './spellValidator';
